+++ NEWS.md
@@ -6,6 +6,7 @@ Red Panda C++ Version 1.0.7
   - fix: When generating project's makefile for clang, don't add -fexec-charset / -finput-charset command line options
   - fix: index of the longest line not correctly updated when inputting with auto completion open
   - enhancement: support UTF-8 BOM files
+  - enhancement: add new tool button for "compiler options"
 
 Red Panda C++ Version 1.0.6
   - fix: gcc compiler set name is not correct in Linux